feat: added queries to DegreeDao and DegreeService
changed var types from int to long

+++ services-interfaces/src/main/java/ar/edu/itba/paw/services/DegreeService.java
@@ -2,5 +2,7 @@ package ar.edu.itba.paw.services;
 
 import ar.edu.itba.paw.models.Degree;
 
-public interface DegreeService extends BaseService<Integer, Degree> {
+
+public interface DegreeService extends BaseService<Long, Degree> {
+
 }
